# Speed up tests that include sleep