The background manager tests do not require 10s to wait for sleep to happen. The pprof code is just utilizing net/http/pprof which we can assume works OK. Any API that requires sampling (cpuprofile, mutex) can not occur in parallel and takes a minimum of 1s.

| test | old time | new time     |
| ---- | -------- | ----------- |
| TestHandlePprofs |  15s  |  0.7s |
| Test_waitForBackgroundManagersToStop | 20s | 0.7s |
| TestOfflineDatabaseStartup | 7.5s | 2.5s |


## Pre-review checklist
- [x] Removed debug logging (`fmt.Print`, `log.Print`, ...)
- [x] Logging sensitive data? Make sure it's tagged (e.g. `base.UD(docID)`, `base.MD(dbName)`)
- [x] Updated relevant information in the API specifications (such as endpoint descriptions, schemas, ...) in `docs/api`

## [Integration Tests](https://jenkins.sgwdev.com/job/SyncGateway-Integration/build?delay=0sec)
- [x] `GSI=true,xattrs=true` https://jenkins.sgwdev.com/job/SyncGateway-Integration/2991/